KTO-826: Move KoulutuksenNimiSection to a separate file
